출력 타입 변

+++ 00. 다시 풀어야 할 문제들/GCD합/GCD합/GCD합.cpp
@@ -33,7 +33,7 @@ int main()
 		{
 			cin >> v[j];
 		}
-		int result = 0;
+		long long result = 0;
 		for (int j= 0; j < n;j++)
 		{
 			for (int k = j+1; k < n; k++)
